+++ packages/next/test/integration/integration-1.test.js
@@ -7,7 +7,7 @@ const builder = require('../../');
 const {
   createRunBuildLambda,
 } = require('../../../../test/lib/run-build-lambda');
-const { duplicateWithConfig } = require('../utils');
+const { duplicateWithConfig } = await import('../utils');
 const { streamToBuffer } = require('@vercel/build-utils');
 const { createHash } = require('crypto');
 
@@ -20,7 +20,7 @@ const SIMPLE_PROJECT = path.resolve(
   '00-middleware'
 );
 
-jest.setTimeout(360000);
+vi.setConfig({ testTimeout: 360000, hookTimeout: 360000 });
 
 function sharedTests(ctx) {
   it('worker uses `middleware` or `middlewarePath` keyword as route path', async () => {